Check link exists before trying to remove

+++ system/version/goos/symlink_darwin.go
@@ -2,6 +2,8 @@ package goos
 
 import (
 	"os"
+
+	"github.com/ironstar-io/tokaido/system/fs"
 )
 
 var tokBinPath = "/usr/local/bin/tok"
@@ -10,14 +12,16 @@ var tokBinPath = "/usr/local/bin/tok"
 func CreateSymlink(path string) error {
 	// Depending on OS, figure out the correct tok path and create symlink
 
-	// Remove any existing soft link
-	err := os.Remove(tokBinPath)
-	if err != nil {
-		return err
+	if fs.CheckExists(tokBinPath) == true {
+		// Remove any existing soft link
+		err := os.Remove(tokBinPath)
+		if err != nil {
+			return err
+		}
 	}
 
 	// Create a new symbolic or "soft" link
-	err = os.Symlink(path, tokBinPath)
+	err := os.Symlink(path, tokBinPath)
 	if err != nil {
 		return err
 	}
