+++ system/version/goos/symlink_windows.go
@@ -3,6 +3,8 @@ package goos
 import (
 	"os"
 	"path/filepath"
+
+	"github.com/ironstar-io/tokaido/system/fs"
 )
 
 // CreateSymlink - Check if a file/folder is writable
@@ -10,13 +12,16 @@ func CreateSymlink(path string) error {
 	// Depending on OS, figure out the correct tok path and create symlink
 	var tokBinPath = filepath.Join("C:/", "Program Files", "Ironstar", "Tokaido", "tok.exe")
 
-	// Remove any existing soft link
-	err := os.Remove(tokBinPath)
-	if err != nil {
-		return err
+	if fs.CheckExists(tokBinPath) == true {
+		// Remove any existing soft link
+		err := os.Remove(tokBinPath)
+		if err != nil {
+			return err
+		}
 	}
+
 	// Create a new symbolic or "soft" link
-	err = os.Symlink(tokBinPath, path)
+	err := os.Symlink(tokBinPath, path)
 	if err != nil {
 		return err
 	}
